Renamed readline_error module as error
